Move canBeVariable table header handling into SortableTable
